WIP changing shortname to sentence
