+++ code/drasil-lang/Language/Drasil.hs
@@ -22,14 +22,15 @@ module Language.Drasil (
   , cross, m2x2, vec2D, dgnl2x2
   -- all the stuff from Unicode
   , Special(..), RenderSpecial(..)
-   -- UID
+  -- UID
   , UID
   -- Classes.Core
   , HasUID(uid)
-  , HasShortName(shortname)
   , HasRefAddress(getRefAdd)
   , HasSymbol(symbol)
   , Referable(..)
+  -- Classes.Core2
+  , HasShortName(shortname)
   -- Classes.Document
   , HasFields(getFields)
   -- Classes
@@ -109,8 +110,6 @@ module Language.Drasil (
   , Sentence(..), SentenceStyle(..), (+:+), (+:+.), (+:), (!.), capSent, ch, sC, sDash, sParen  
   -- Sentence.Extract
   , sdep, shortdep
-  -- RefProg
-  , Reference(..), RefInfo(..)
   -- NounPhrase
   , NounPhrase(..), NP, pn, pn', pn'', pn''', pnIrr, cn, cn', cn'', cn''', cnIP
   , cnIrr, cnIES, cnICES, cnIS, cnUM, nounPhrase, nounPhrase'
@@ -146,8 +145,10 @@ module Language.Drasil (
   -- Symbol.Helpers
   , eqSymb, codeSymb, hasStageSymbol
   , autoStage, hat, prime, staged, sub, subStr, sup , unicodeConv, upperLeft, vec
+  -- RefProg
+  , RefInfo(..)
   -- Reference
-  , makeRef2S, makeCite, makeCiteS, makeRef2, makeCiteInfo, makeCiteInfoS, rw
+  , Reference(..), makeRef2S, makeCite, makeCiteS, makeRef2, makeCiteInfo, makeCiteInfoS, rw
   -- Label.Type
   , getAdd, prepend
   , LblType(RP, Citation, URI), IRefProg(..)
@@ -215,7 +216,8 @@ import Language.Drasil.Document.Core (Contents(..), ListType(..), ItemType(..), 
 import Language.Drasil.Unicode -- all of it
 import Language.Drasil.UID (UID)
 import Language.Drasil.Classes.Core (HasUID(uid), HasSymbol(symbol),
-  HasRefAddress(getRefAdd), HasShortName(shortname), Referable(refAdd, renderRef))
+  HasRefAddress(getRefAdd), Referable(refAdd, renderRef))
+import Language.Drasil.Classes.Core2 (HasShortName(shortname))
 import Language.Drasil.Classes (NamedIdea(term), Idea(getA),
   Definition(defn), ConceptDomain(cdom), Concept, HasUnitSymbol(usymb),
   IsUnit(getUnits), CommonIdea(abrv), HasAdditionalNotes(getNotes), Constrained(constraints), 
@@ -275,7 +277,8 @@ import Language.Drasil.Space (Space(..), RealInterval(..), Inclusive(..),
 import Language.Drasil.Sentence (Sentence(..), SentenceStyle(..), (+:+),
   (+:+.), (+:), (!.), capSent, ch, sC, sDash, sParen)
 import Language.Drasil.Sentence.Extract (sdep, shortdep) -- exported for drasil-database FIXME: move to development package?
-import Language.Drasil.Reference (makeCite, makeCiteS, makeRef2, makeRef2S, makeCiteInfo, makeCiteInfoS, rw)
+import Language.Drasil.RefProg (RefInfo(..))
+import Language.Drasil.Reference (Reference(..), makeCite, makeCiteS, makeRef2, makeRef2S, makeCiteInfo, makeCiteInfoS, rw)
 import Language.Drasil.Symbol (Decoration(..), Symbol(..), compsy)
 import Language.Drasil.Symbol.Helpers (eqSymb, codeSymb, hasStageSymbol,
   autoStage, hat, prime, staged, sub, subStr, sup, unicodeConv, upperLeft, vec)
@@ -284,7 +287,6 @@ import Language.Drasil.Misc -- all of it
 import Language.Drasil.People (People, Person, person, HasName(..),
   person', personWM, personWM', mononym, name, nameStr, rendPersLFM, 
   rendPersLFM', rendPersLFM'', comparePeople)
-import Language.Drasil.RefProg (Reference(..), RefInfo(..))
 import Language.Drasil.Label.Type (getAdd, LblType(RP, Citation, URI), IRefProg(..), prepend)
 
 import Language.Drasil.UnitLang (USymb(US))
