+++ code/drasil-lang/Language/Drasil/Classes.hs
@@ -36,7 +36,7 @@ import Language.Drasil.Derivation (Derivation)
 import Language.Drasil.UnitLang(UDefn, USymb)
 import Language.Drasil.Expr (Expr, Relation)
 import Language.Drasil.NounPhrase.Core (NP)
-import Language.Drasil.RefProg (Reference)
+import Language.Drasil.Reference (Reference)
 import Language.Drasil.Space (Space)
 import Language.Drasil.Sentence (Sentence)
 import Language.Drasil.UID (UID)
