+++ code/drasil-lang/Language/Drasil/Classes/Core.hs
@@ -2,13 +2,11 @@
 -- | Defining the core classes which represent knowledge-about-knowledge
 module Language.Drasil.Classes.Core (
     HasUID(uid)
-  , HasShortName(shortname)
   , HasRefAddress(getRefAdd)
   , HasSymbol(symbol)
   , Referable(refAdd, renderRef)
   ) where
 
-import Language.Drasil.ShortName (ShortName)
 import Language.Drasil.Stages (Stage)
 import Language.Drasil.Symbol (Symbol)
 import Language.Drasil.UID (UID)
@@ -20,12 +18,6 @@ import Control.Lens (Lens')
 class HasUID c where
   -- | Provides a /unique/ id for internal Drasil use.
   uid :: Lens' c UID
-
--- | A 'ShortName' is the text to be displayed for a link.
--- Used for referencing within a document that can include symbols and whatnot if required.
--- Visible in the typeset documents (pdf).
-class HasShortName  s where
-  shortname :: s -> ShortName
 
 -- | A HasSymbol is anything which has a 'Symbol'.
 class HasSymbol c where
