+++ code/drasil-lang/Language/Drasil/ShortName.hs
@@ -1,12 +1,14 @@
 module Language.Drasil.ShortName where
 
+import Language.Drasil.Sentence
+
 -- | Used for holding the short form of a name (as a String with a wrapper).
-newtype ShortName = ShortNm String
+newtype ShortName = ShortNm Sentence
 
 -- | Pulls the short form (as a 'String') out of a 'ShortName'.
-getStringSN :: ShortName -> String
+getStringSN :: ShortName -> Sentence
 getStringSN (ShortNm s) = s
 
 -- | Smart constructor for making a 'String' into a 'ShortName'.
-shortname' :: String -> ShortName
+shortname' :: Sentence -> ShortName
 shortname' = ShortNm
